Renamed AcmeDemoBundle to DemoBundle
