--- Resources/skeleton/acme-demo-bundle/Acme/DemoBundle/AcmeDemoBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Acme\DemoBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class AcmeDemoBundle extends Bundle
-{
-}
