--- Resources/skeleton/acme-demo-bundle/Acme/DemoBundle/Controller/WelcomeController.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Acme\DemoBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-class WelcomeController extends Controller
-{
-    public function indexAction()
-    {
-        /*
-         * The action's view can be rendered using render() method
-         * or @Template annotation as demonstrated in DemoController.
-         *
-         */
-
-        return $this->render('AcmeDemoBundle:Welcome:index.html.twig');
-    }
-}
