--- Resources/skeleton/acme-demo-bundle/Acme/DemoBundle/Form/ContactType.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Acme\DemoBundle\Form;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-
-class ContactType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder->add('email', 'email');
-        $builder->add('message', 'textarea');
-    }
-
-    public function getName()
-    {
-        return 'contact';
-    }
-}
