--- Resources/skeleton/acme-demo-bundle/Acme/DemoBundle/Resources/views/Demo/hello.html.twig
@@ -1,9 +1,0 @@
-{% extends "AcmeDemoBundle::layout.html.twig" %}
-
-{% block title "Hello " ~ name %}
-
-{% block content %}
-    <h1>Hello {{ name }}!</h1>
-{% endblock %}
-
-{% set code = code(_self) %}
