+++ Resources/skeleton/demo-bundle/DemoBundle/Controller/WelcomeController.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace DemoBundle\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+
+class WelcomeController extends Controller
+{
+    public function indexAction()
+    {
+        /*
+         * The action's view can be rendered using render() method
+         * or @Template annotation as demonstrated in DemoController.
+         *
+         */
+
+        return $this->render('DemoBundle:Welcome:index.html.twig');
+    }
+}
