+++ Resources/skeleton/demo-bundle/DemoBundle/DemoBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace DemoBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class DemoBundle extends Bundle
+{
+}
