+++ Resources/skeleton/demo-bundle/DemoBundle/Resources/public/images/field-background.gif
