+++ Resources/skeleton/demo-bundle/DemoBundle/Resources/views/Demo/hello.html.twig
@@ -1,0 +1,9 @@
+{% extends "DemoBundle::layout.html.twig" %}
+
+{% block title "Hello " ~ name %}
+
+{% block content %}
+    <h1>Hello {{ name }}!</h1>
+{% endblock %}
+
+{% set code = code(_self) %}
