+++ Resources/skeleton/demo-bundle/DemoBundle/Resources/views/Secured/helloadmin.html.twig
@@ -1,0 +1,9 @@
+{% extends "DemoBundle:Secured:layout.html.twig" %}
+
+{% block title "Hello " ~ name %}
+
+{% block content %}
+    <h1 class="title">Hello {{ name }} secured for Admins only!</h1>
+{% endblock %}
+
+{% set code = code(_self) %}
